Moved media to util. Also resolving issues 3, 11 and 15 because I fixed them earlier and forgot to close them.
Fixes #3
Fixes #11
Fixes #15
