+++ app/src/main/java/com/luorrak/ouroboros/deepzoom/DeepZoomFragment.java
@@ -16,7 +16,7 @@ import android.widget.Toast;
 import com.koushikdutta.async.future.FutureCallback;
 import com.koushikdutta.ion.Ion;
 import com.luorrak.ouroboros.R;
-import com.luorrak.ouroboros.gallery.Media;
+import com.luorrak.ouroboros.util.Media;
 import com.luorrak.ouroboros.util.ChanUrls;
 import com.luorrak.ouroboros.util.InfiniteDbHelper;
 import com.luorrak.ouroboros.util.NetworkHelper;
