+++ app/src/main/java/com/luorrak/ouroboros/gallery/GalleryFragment.java
@@ -15,6 +15,7 @@ import android.view.View;
 import android.view.ViewGroup;
 
 import com.luorrak.ouroboros.R;
+import com.luorrak.ouroboros.util.Media;
 import com.luorrak.ouroboros.util.DbContract;
 import com.luorrak.ouroboros.util.InfiniteDbHelper;
 import com.luorrak.ouroboros.util.NetworkHelper;
